Add poetry support and update python project dependencies in pyproject.toml (#1)
* Add poetry support and update python project dependencies in pyproject.toml

Added new python libraries and updated their versions in the pyproject.toml file. This update ensures compatibility with the latest versions of these libraries and also improves the overall performance and security of the application.

Additionally, added a new poetry.lock file to lock the dependency versions. This provides consistent installs across machines, ideal for a production environment.

This update also addresses the following:
- Upgraded aiohttp from 3.8.0 to 3.9.0 for better performance.
- Added pytest and black for better testing and formatting respectively under the dev dependencies.

* Add asyncio_mode to pytest configuration

This commit adds a new configuration line to the pytest.ini file: "asyncio_mode = auto". This is necessary to configure the default mode for the pytest-asyncio plugin - in this case, set to "auto" to automatically detect asyncio tests. This change should help to improve the detection and execution of asynchronous test cases.